// src/app/admin/schedule/schedule.enums.ts
// Define Schedule Module Enums below

export enum ScheduleModuleEnum {
  AllCandidateInviteStatus = 'All',
  DeliveredCandidateInviteStatus = 'delivered',
  FailedCandidateInviteStatus = 'failed',
  CreatingScheduleAssessmentStatus = 'Scheduling For New Assessment...',
  CreatedScheduleAssessmentStatus = 'New Schedule Created For Assessment',
  FailedScheduleAssessmentStatus = 'Failed To Schedule Assessment',
  CsvUpload = 'csv',
  CustomUpload = 'custom',
  EmailStatus = 'EmailStatus',
  ScheduleStatus = 'ScheduleStatus'
}